# ui/src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "get": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/callback": {
      "get": {
        "operationId": "AuthController_callback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/user": {
      "get": {
        "operationId": "AuthController_getUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user information",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout current user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/cases/{id}": {
      "get": {
        "operationId": "CasesController_getCaseById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Case ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interview case details"
          }
        },
        "summary": "Get a specific interview case by ID",
        "tags": [
          "cases"
        ]
      }
    }
  },
  "info": {
    "title": "System Design Interview Simulator API",
    "description": "API for the SD Interview Simulator application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "subscriptionStatus": {
            "type": "string"
          },
          "interviewsCompleted": {
            "type": "number"
          },
          "interviewsRemaining": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "avatarUrl",
          "subscriptionStatus",
          "interviewsCompleted",
          "interviewsRemaining"
        ]
      },
...
    }
  }
}